fix clippy: remove stale assignment-phase fields and minor lints [sp-ryve-f27d981b]
- Remove assignment_phase/phase_* fields from test HandAssignment constructors
  (these fields no longer exist after reverting the assignment epic)
- Remove source_branch/target_branch from NewHandAssignment in test
- Collapse nested ifs per clippy::collapsible_if
- Remove unnecessary i32 casts and vec! in tests
- Wire up recompute_spark_summary, recompute_hand_counts,
  rebuild_precomputed_maps, recompute_git_stats calls in app.rs

+++ src/screen/sparks.rs
@@ -2457,7 +2457,7 @@ mod tests {
                 title: format!("Epic {e}"),
                 spark_type: "epic".to_string(),
                 status: "open".to_string(),
-                priority: (e % 5) as i32,
+                priority: e % 5,
                 parent_id: None,
                 ..mk_sort_spark(&eid, 0, "epic", "open")
             });
@@ -2468,7 +2468,7 @@ mod tests {
                     title: format!("Task {e}-{c}"),
                     spark_type: "task".to_string(),
                     status: "open".to_string(),
-                    priority: (c % 5) as i32,
+                    priority: c % 5,
                     parent_id: Some(eid.clone()),
                     assignee: if c % 3 == 0 {
                         Some(format!("agent-{c}"))
